More children are born!
The model now includes recurrent births with fertility evaluated at the beginning of every year for those not already scheduled to give birth.  The model also includes breastfeeding so that women cannot give birth straight after having a child.  DUration of breastfeeding is currently set at 3 months, but can be changed as needed by changing the value in just one place.

Potential limitations:
Fertility is assigned between age and max age range.  So some women can give birth in 4 years if they are for example 15.  Might want to change that if birth rates are too low e.g. min=Age, and max=Age+9months

+++ person.h
@@ -23,7 +23,7 @@ public:
 	int Sex;
 
 
-	int DoB;										// Variables related to age
+	double DoB;										// Variables related to age
 	int AgeT0;
 	double Age;
